requirements.txt
amqp==2.6.1
anyjson==0.3.3
appdirs==1.4.4
asgiref==3.2.7
astroid==2.4.1
attrs==20.1.0
beautifulsoup4==4.9.1
billiard==3.6.3.0
black==19.10b0
boto3==1.15.6
botocore==1.18.6
Brotli==1.0.7
celery==4.4.7
certifi==2020.4.5.1
cffi==1.14.1
cfgv==3.1.0
chardet==3.0.4
click==7.1.2
colorama==0.4.3
cryptography==3.0
dash==1.10.0
dash-bootstrap-components==0.10.3
dash-core-components==1.9.0
dash-daq==0.5.0
dash-html-components==1.0.3
dash-renderer==1.3.0
dash-table==4.6.2
distlib==0.3.0
Django==3.0.6
django-bootstrap4==2.2.0
django-celery-results==1.2.1
django-plotly-dash==1.3.1
djangorestframework==3.11.0
djangorestframework-datatables==0.5.2
dnspython==1.16.0
dpd-components==0.1.0
dpd-static-support==0.0.5
eventlet==0.26.1
filelock==3.0.12
Flask==1.1.2
Flask-Compress==1.5.0
future==0.18.2
geographiclib==1.50
geopy==2.0.0
googlemaps==4.4.2
greenlet==0.4.16
gunicorn==20.0.4
identify==1.4.15
idna==2.9
importlib-metadata==1.6.0
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
jmespath==0.10.0
kombu==4.6.11
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
monotonic==1.5
nodeenv==1.3.5
numpy==1.18.4
pandas==1.0.3
pathspec==0.8.0
plotly==4.7.1
pre-commit==2.3.0
psycopg2==2.8.6
pycparser==2.20
pylint==2.5.2
pyOpenSSL==19.1.0
python-dateutil==2.8.1
python-dotenv==0.14.0
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
regex==2020.7.14
requests==2.24.0
retrying==1.3.3
s3transfer==0.3.3
selenium==3.141.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
toml==0.10.1
typed-ast==1.4.1
urllib3==1.25.9
vine==1.3.0
virtualenv==20.0.20
Werkzeug==1.0.1
whitenoise==5.1.0
wrapt==1.12.1
xmltodict==0.12.0
zipp==3.1.0
zope.event==4.4
zope.interface==5.1.0
